+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `#[allow(clippy::all)]` in generated code
 - `is_empty` functions on maps
+- impl `Ordinal` for arrays
 
 # 0.1.7 - 2025-06-02